indicator_indicator many-to-many relation added

+++ src/main/java/lab/zhang/hermes/dao/IndicatorDao.java
@@ -15,7 +15,7 @@ public interface IndicatorDao {
 
     IndicatorEntity findOne(long id);
 
-    Long insert(IndicatorEntity entity);
+    int insert(IndicatorEntity entity);
 
     void update(IndicatorEntity entity);
