+++ src/test/java/unit/lab/zhang/hermes/dao/IndicatorDaoTest.java
@@ -47,7 +47,7 @@ public class IndicatorDaoTest {
     @Test
     public void test_insert() {
         IndicatorEntity item = new IndicatorEntity("测试式3", "{\"name\":\"指标1\",\"type\":1,\"id\":0,\"value\":1}");
-        Long id = target.insert(item);
+        int id = target.insert(item);
         assertNotNull(id);
         assertTrue(id > 0);
         System.out.println(id);
